+++ src/types/Number.ts
@@ -1,0 +1,7 @@
+export type NumericParser = {
+  parse: (str: string) => number | typeof NaN;
+};
+
+export type NumericFormatter = {
+  format: (num: number) => string;
+};
